SPT-0011 - Review validate.py and consolidate remaining files (#12)
* fix main branch errors

* consolidate functions

* review validate functions

remove specific checks

* remove prediction_checker references

* fix unit test import

* consolidate unit tests
