+++ cishouseholds/impute.py
@@ -8,7 +8,9 @@ from typing import Union
 
 from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
+from pyspark.sql.types import ArrayType
 from pyspark.sql.types import DoubleType
+from pyspark.sql.types import StringType
 from pyspark.sql.window import Window
 
 from cishouseholds.derive import assign_random_day_in_month
@@ -17,7 +19,19 @@ from cishouseholds.merge import union_multiple_tables
 from cishouseholds.pipeline.load import extract_from_table
 from cishouseholds.pipeline.load import update_table
 from cishouseholds.pyspark_utils import get_or_create_spark_session
-from cishouseholds.udfs import generate_sample_proportional_to_size_udf
+
+
+def generate_sample_proportional_to_size_udf(spark):
+    def sample_proportional_to_size(imputation_variables, expected_frequencies, number):
+        """UDF for sampling proportional to size."""
+        import numpy as np
+
+        results = np.random.choice(a=imputation_variables, p=expected_frequencies, replace=False, size=number).tolist()
+        return results
+
+    spark.udf.register("sample_proportional_to_size_udf_reg", sample_proportional_to_size)
+    return F.udf(sample_proportional_to_size, ArrayType(StringType()))
+
 
 sample_proportional_to_size_udf = generate_sample_proportional_to_size_udf(get_or_create_spark_session())
